CRDT Op introduction of subcrate
