+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "3"
-members = ["xaeroflux", "xaeroflux-actors", "xaeroflux-core", "xaeroflux-ffi", "xaeroflux-macros"]
+members = ["xaeroflux", "xaeroflux-actors", "xaeroflux-core", "xaeroflux-crdts", "xaeroflux-ffi", "xaeroflux-macros"]
