+++ xaeroflux-actors/Cargo.toml
@@ -33,8 +33,9 @@ serial_test = "0.10.0"
 libc = "0.2.172"
 
 [dependencies]
-xaeroflux-core = {version = "0.5.0-m3"}
+xaeroflux-core = {path = "../xaeroflux-core"}
 xaeroflux-macros = {path = "../xaeroflux-macros"}
+xaeroflux-crdts = {path = "../xaeroflux-crdts"}
 ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
 rand = "0.8"
 iroh = {version = "0.35.0", "default-features" = true, features = ["discovery-local-network", "discovery-pkarr-dht"] }
