+++ xaeroflux-actors/src/aof/actor.rs
@@ -111,7 +111,8 @@ mod tests {
     };
 
     use super::*;
-    use crate::{XaeroEvent, subject::SubjectHash};
+    use crate::subject::SubjectHash;
+    use xaeroflux_core::event::XaeroEvent;
 
     /// Before each test, override CONF.aof.file_path to point at a temporary directory,
     /// so that `emit_control_path_with_subject_hash(...)` always has a valid base path.
