--- xaeroflux-actors/src/crdt/actors/mod.rs
@@ -1,2 +1,0 @@
-pub mod crdt_op;
-pub mod bus;
