--- xaeroflux-actors/src/crdt/bus.rs
@@ -1,6 +1,0 @@
-/// CRDT Bus for sending and receiving CRDT operations
-pub struct CRDTBus {
-    pub tx: Sender<CrdtOp>,
-    pub rx: Receiver<CrdtOp>,
-}
-
