--- xaeroflux-actors/src/crdt/mod.rs
@@ -1,3 +1,0 @@
-pub mod actors;
-pub mod core;
-pub mod bus;
