+++ xaeroflux-actors/src/lib.rs
@@ -27,39 +27,7 @@ static NEXT_ID: AtomicU64 = AtomicU64::new(1);
 fn next_id() -> u64 {
     NEXT_ID.fetch_add(1, Ordering::SeqCst)
 }
-/// Envelope wrapping an application or system `Event` payload
-/// along with an optional Merkle inclusion proof.
-#[derive(Debug, Clone)]
-pub struct XaeroEvent {
-    /// Core event data (e.g., domain event encoded as bytes).
-    pub evt: Event<Vec<u8>>,
-    /// Optional Merkle proof bytes (e.g., from MMR).
-    pub merkle_proof: Option<Vec<u8>>,
-}
-
-#[repr(C)]
-#[derive(Clone, Copy)]
-pub struct ScanWindow {
-    pub start: u64,
-    pub end: u64,
-}
-unsafe impl Pod for ScanWindow {}
-unsafe impl Zeroable for ScanWindow {}
-
-/// Defines per-event pipeline operations.
-#[derive(Clone)]
-pub enum Operator {
-    Scan(Arc<ScanWindow>),
-    /// Transform the event into another event.
-    Map(Arc<dyn Fn(XaeroEvent) -> XaeroEvent + Send + Sync>),
-    /// Keep only events matching the predicate.
-    Filter(Arc<dyn Fn(&XaeroEvent) -> bool + Send + Sync>),
-    /// Drop events without a Merkle proof.
-    FilterMerkleProofs,
-    /// Terminal op: drop all events.
-    Blackhole,
-}
-
+use xaeroflux_core::event::{ScanWindow, XaeroEvent};
 /// Holds your system actors and subscription to keep them alive.
 pub struct XFluxHandle {
     /// Composite subscription for core events.
