+++ xaeroflux-crdts/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "xaeroflux"
+name = "xaeroflux-crdts"
 version = "0.5.0-m3"
 edition     = "2024"
-description = "EXPERIMENTAL: actor layer for xaeroflux"
+description = "EXPERIMENTAL: CRDT layer for xaeroflux"
 categories = ["database", "database-implementations"]
 keywords = ["unstable", "draft", "experimental"]
 authors = ["ricky.nj@gmail.com"]
@@ -12,8 +12,16 @@ repository = "https://github.com/block-xaero/xaeroflux.git"
 homepage = "https://blockxaero.io"
 readme = "README.md"
 
+[lib]
+name = "xaeroflux_crdt"
+crate-type = ["rlib"]
+
 [dependencies]
-xaeroflux-core = "0.5.0-m3"
-xaeroflux-actors =  "0.5.0-m3"
-xaeroflux-macros =  "0.5.0-m3"
-xaeroflux-crdts = {path = "../xaeroflux-crdts"}
+# Core CRDT algorithms
+crdts = "7.3.2"
+
+# Zero-copy serialization (same as xaeroflux-core)
+rkyv = "0.8.10"
+bytemuck = "1.23.0"
+serde = { version = "1.0.219", features = ["derive"] }
+
